Convert script to test; rename / delete others
